Completed tasks 4,5 and 6
finished adding the background mario level and adding controls to move the viewport left and right, limited to the size of the background image.

Created a cursor class that updates a sprite to the position of the cursor.

Added the J cole meal enemy for task 6 as my wifi is wonky at the moment and thats the best thing i had in my downloads folder.